Master v5

See merge request trondjro/hverdagshelt---team-3!231

+++ server/src/routes/events.js
@@ -1,5 +1,7 @@
 //@flow
-import {Event} from '../models';
+import { Event, User } from '../models';
+import { tokenManager } from '../tokenManager';
+import * as passwordHash from '../passwordHash';
 
 type Request = express$Request;
 type Response = express$Response;
@@ -8,50 +10,61 @@ const app = require('../app');
 
 //Events
 app.get('/municipals/:id/events', (req: Request, res: Response) => {
-    return Event.findAll({ where: { munId: Number(req.params.id) },
-                            order: [['timeStart', 'ASC']]})
-      .then(events => events ? res.send(events) : res.sendStatus(404)
-    );
+  return Event.findAll({ where: { munId: Number(req.params.id) }, order: [['timeStart', 'ASC']] }).then(events =>
+    events ? res.send(events) : res.sendStatus(404)
+  );
 });
 
 //Event
-app.get('/secure/events', (req: Request, res: Response) => {
-    return Event.findAll().then(events => res.send(events));
+app.get('/events', (req: Request, res: Response) => {
+  return Event.findAll().then(events => res.send(events));
 });
-app.get('/secure/events/:id', (req: Request, res: Response) => {
-    return Event.findOne({ where: { eventId: Number(req.params.id) } }).then(event =>
-        event ? res.send(event) : res.sendStatus(404)
-    );
+app.get('/events/:id', (req: Request, res: Response) => {
+  return Event.findOne({ where: { eventId: Number(req.params.id) } }).then(event =>
+    event ? res.send(event) : res.sendStatus(404)
+  );
 });
-app.put('/secure/events/:id', (req: Request, res: Response) => {
-    //Flow type checking mixed src: https://github.com/flow-typed/flow-typed/issues/812
-    const body = req.body !== null && typeof req.body === 'object' ? req.body : {};
-    const { title, content, image, longitude, latitude, timeStart, timeEnd } = body;
-
-    return Event.update(
+app.put('/events/:id', (req: Request, res: Response) => {
+  //Flow type checking mixed src: https://github.com/flow-typed/flow-typed/issues/812
+  const body = req.body !== null && typeof req.body === 'object' ? req.body : {};
+  const { title, content, image, longitude, latitude, timeStart, timeEnd } = body;
+  console.log(image);
+  let tokenData = tokenManager.verifyToken(req.headers['x-access-token']);
+  if (tokenData) {
+    if (tokenData.rank === 4 || tokenData.rank === 3) {
+      return Event.update(
         {
-            title: title,
-            content: content,
-            image: image,
-            longitude: longitude,
-            latitude: latitude,
-            timeStart: timeStart,
-            timeEnd: timeEnd
+          title: title,
+          content: content,
+          image: image,
+          longitude: longitude,
+          latitude: latitude,
+          timeStart: timeStart,
+          timeEnd: timeEnd
         },
         {
-            where: {
-                eventId: req.params.id
-            }
+          where: {
+            eventId: req.params.id
+          }
         }
-    ).then(count => (count ? res.sendStatus(200) : res.sendStatus(404)));
+      ).then(count => (count ? res.sendStatus(200) : res.sendStatus(404)));
+    } else {
+      res.sendStatus(401);
+    }
+  } else {
+    res.sendStatus(401);
+  }
 });
-app.post('/secure/events', (req: Request, res: Response) => {
-    // if (!(req.body instanceof Object)) return res.sendStatus(400);
-    //Flow type checking mixed src: https://github.com/flow-typed/flow-typed/issues/812
-    const body = req.body !== null && typeof req.body === 'object' ? req.body : {};
-    const { title, content, image, longitude, latitude, timeStart, timeEnd, categoryId, munId, userId } = body;
+app.post('/events', (req: Request, res: Response) => {
+  // if (!(req.body instanceof Object)) return res.sendStatus(400);
+  //Flow type checking mixed src: https://github.com/flow-typed/flow-typed/issues/812
+  const body = req.body !== null && typeof req.body === 'object' ? req.body : {};
+  const { title, content, image, longitude, latitude, timeStart, timeEnd, categoryId, munId, userId } = body;
 
-    return Event.create({
+  let tokenData = tokenManager.verifyToken(req.headers['x-access-token']);
+  if (tokenData) {
+    if (tokenData.rank === 4 || tokenData.rank === 3) {
+      return Event.create({
         title: title,
         content: content,
         image: image,
@@ -62,12 +75,28 @@ app.post('/secure/events', (req: Request, res: Response) => {
         categoryId: categoryId,
         munId: munId,
         userId: userId
-    }).then(count => (count ? res.sendStatus(200) : res.sendStatus(404)));
+      }).then(count => (count ? res.send(count) : res.sendStatus(404)));
+    } else {
+      res.sendStatus(401);
+    }
+  } else {
+    res.sendStatus(401);
+  }
 });
-app.delete('/secure/events/:id', (req: Request, res: Response) => {
-    return Event.destroy({
+
+app.delete('/events/:id', (req: Request, res: Response) => {
+  let tokenData = tokenManager.verifyToken(req.headers['x-access-token']);
+  if (tokenData) {
+    if (tokenData.rank === 4 || tokenData.rank === 3) {
+      return Event.destroy({
         where: {
-            eventId: req.params.id
+          eventId: req.params.id
         }
-    }).then(count => (count ? res.sendStatus(200) : res.sendStatus(404)));
+      }).then(count => (count ? res.sendStatus(200) : res.sendStatus(404)));
+    } else {
+      res.sendStatus(401);
+    }
+  } else {
+    res.sendStatus(401);
+  }
 });
